Implement video_call_request_page
